use int64 instead

+++ propstoken/client.go
@@ -11,8 +11,8 @@ import (
 type DelegatedTransfer struct {
 	Signature string `json:"signature" gencodec:"required"`
 	To        string `json:"to" gencodec:"required"`
-	Amount    int    `json:"amount" gencodec:"required"`
-	Fee       int    `json:"fee" gencodec:"required"`
+	Amount    int64  `json:"amount" gencodec:"required"`
+	Fee       int64  `json:"fee" gencodec:"required"`
 	Nonce     string `json:"nonce"    gencodec:"required"`
 	From      string `json:"from" gencodec:"required"`
 }
